changed the db functions and mongo db setup

+++ data_access_layer/basic_db_functions.py
@@ -1,22 +1,36 @@
 import random
-from pymongo import MongoClient
-from db.connection_credentials import mongodb_url
 
-client = MongoClient(mongodb_url)
-database = client['telegram_bot_data']
-collection = database['tests']
+from globals import globals
 
 def test_store_in_db():
-    random_number = random.randint(1, 1000)
-    if not check_exists_in_db(number=random_number):
-        collection.insert_one({'random_number': random_number})
-    return random_number
+    mongo_client = globals.mongo_client
+    if mongo_client:
+        db = mongo_client["test_db"]
+        collection = db["test"]
+        random_number = random.randint(1, 1000)
+        if not check_exists_in_db(number=random_number):
+            collection.insert_one({'random_number': random_number})
+        return random_number
 
 def check_exists_in_db(number: int):
-    return collection.find_one({'random_number': number})
+    mongo_client = globals.mongo_client
+    if mongo_client:
+        db = mongo_client["test_db"]
+        collection = db["test"]
+        return collection.find_one({'random_number': number})
+    else:
+        raise ValueError("mongo db client is not set up")
 
 def delete_number_from_db(number: int):
-    if check_exists_in_db(number=number):
-        collection.delete_one({'random_number': number})
-        return True
-    return False
+    try:
+        mongo_client = globals.mongo_client
+        if mongo_client:
+            db = mongo_client["test_db"]
+            collection = db["test"]
+            if check_exists_in_db(number=number):
+                collection.delete_one({'random_number': number})
+                return True
+            return False
+    except Exception as e:
+        print(e)
+        return e
